requirements.txt
requests
BeautifulSoup4
tenacity
regex
Pillow